Change signatures to c_int and c_uint.

+++ ailets-rs/actor_runtime/src/actor_runtime.rs
@@ -1,11 +1,11 @@
-use std::os::raw::c_char;
+use std::os::raw::{c_char, c_int, c_uint};
 
 #[link(wasm_import_module = "")]
 extern "C" {
-    pub fn n_of_streams(name_ptr: *const c_char) -> i32;
-    pub fn open_read(name_ptr: *const c_char, index: usize) -> i32;
-    pub fn open_write(name_ptr: *const c_char) -> i32;
-    pub fn aread(fd: i32, buffer_ptr: *mut u8, count: usize) -> i32;
-    pub fn awrite(fd: i32, buffer_ptr: *const u8, count: usize) -> i32;
-    pub fn aclose(fd: i32) -> i32;
+    pub fn n_of_streams(name_ptr: *const c_char) -> c_int;
+    pub fn open_read(name_ptr: *const c_char, index: c_uint) -> c_int;
+    pub fn open_write(name_ptr: *const c_char) -> c_int;
+    pub fn aread(fd: c_int, buffer_ptr: *mut u8, count: c_uint) -> c_int;
+    pub fn awrite(fd: c_int, buffer_ptr: *const u8, count: c_uint) -> c_int;
+    pub fn aclose(fd: c_int) -> c_int;
 }
